+++ src/html/tags.rs
@@ -230,12 +230,15 @@ declare_common_tags_and_macro! {
     template;
 }
 
+declare_tags! {
+    style;  //  conflicts with html::attributes::style, attribute::style    > tags::style
+}
+
 // These are non-common tags
 // which the users need to explicitly import using
 // html::tags::style, html::tags::html, etc
 //
 declare_tags_and_macro! {
-    style;  //  conflicts with html::attributes::style, attribute::style    > tags::style
     title; // conflicts with html::attributes::title  , attributes::title   > tags::title
     slot;  // conflicts with html::attributes::slot   , attrributes::slot   > tags::slot
 }