+++ src/lib.rs
@@ -1,4 +1,4 @@
-//#![deny(warnings)]
+#![deny(warnings)]
 #![deny(clippy::all)]
 #![feature(arbitrary_self_types)]
 #![doc(
@@ -215,6 +215,7 @@ pub use sauron_vdom::{
     Callback,
     Dispatch,
     Text,
+    Value,
 };
 
 /// A simplified version of saurdon_vdom node, where we supplied the type for the tag